Move middleware to bin/glusterd2/middleware
