+++ bin/glusterd2/middleware/auth.go
