+++ bin/glusterd2/middleware/recover.go
